perbaikan pada, RBAC, dan penggabungan admin, user, dan karyawan controller

+++ routes/api.php
@@ -2,25 +2,22 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\AuthController;
-use App\Http\Controllers\AdminController;
+use App\Http\Controllers\UsersController;
 use App\Http\Controllers\BarangController;
-
-
 
 Route::group([
     'middleware' => 'api',
     'prefix' => 'auth',
 ], function () {
-    Route::post('masuk', [AuthController::class, 'masuk']);
-    Route::post('daftar', [AuthController::class, 'daftar']);
-    Route::get('users', [AuthController::class, 'listuser']);
-    Route::delete('user/{id}', [AuthController::class, 'hapus']);
-    //bagian admin
-    Route::get('admins' ,[AdminController::class, 'listAdmin']);
-    Route::post('admin/daftar', [AdminController::class, 'daftarAdmin']);
-    Route::delete('admin/delete/{id}', [AdminController::class, 'hapusAdmin']);
-    Route::post('admin/login', [AdminController::class, 'masukAdmin']);
+    Route::post('masuk', [UsersController::class, 'masuk']);
+    Route::post('daftar', [UsersController::class, 'daftar']);
+    Route::get('users', [UsersController::class, 'listUsersByRole'])->middleware('role:admin');
+    Route::delete('users/{role}/{id}', [UsersController::class, 'hapusByRole'])->middleware('role:admin');
+    Route::post('admin/daftar', [UsersController::class, 'daftarAdmin'])->middleware('role:admin');
+
+    Route::post('keluar', [UsersController::class, 'keluar'])->middleware('auth:api');
+    Route::post('refresh', [UsersController::class, 'refresh'])->middleware('auth:api');
+    Route::get('saya', [UsersController::class, 'saya'])->middleware('auth:api');
 });
 
 Route::group([
@@ -28,6 +25,7 @@ Route::group([
     'prefix' => 'produk',
 ], function () {
     Route::get('barang-list', [BarangController::class, 'listBarang']);
-    Route::post('barang-tambah', [BarangController::class, 'tambahbarang']);
-    Route::delete('barang-hapus/{id}', [BarangController::class, 'hapusbarang']);
+    Route::post('barang-tambah', [BarangController::class, 'tambahbarang'])->middleware('role:admin,karyawan');
+    Route::post('barang-edit/{id}', [BarangController::class, 'editbarang'])->middleware('role:admin,karyawan');
+    Route::delete('barang-hapus/{id}', [BarangController::class, 'hapusbarang'])->middleware('role:admin,karyawan');
 });
